+++ apps/web/src/DashboardReminder.tsx
@@ -17,7 +17,6 @@ import { Card, CardContent } from "./components/ui/card"
 import { useConnection } from "./providers/ConnectionProvider"
 import { getField, getFieldCustomization, isRelationField, tryFunction } from "@stoker-platform/utils"
 import { getFormattedFieldValue } from "./utils/getFormattedFieldValue"
-import { getSortingValue } from "./utils/getSortingValue"
 
 interface DashboardReminderProps {
     reminder: Reminder
@@ -48,6 +47,10 @@ export const DashboardReminder = ({ reminder, title, collection }: DashboardRemi
     const [sorting, setSorting] = useState<{ field: string; direction: "asc" | "desc" } | undefined>(undefined)
     const [rowHighlight, setRowHighlight] = useState<RowHighlight[] | undefined>(undefined)
     const pages = 5
+
+    const setResultsCallback = useCallback((results: StokerRecord[]) => {
+        setResults(reminder.filter ? results.filter((result) => reminder.filter?.(result)) : results)
+    }, [])
 
     const loadingTimeoutRef = useRef<NodeJS.Timeout | null>(null)
     const debouncedSetIsLoading = useCallback((loading: boolean) => {
@@ -96,7 +99,14 @@ export const DashboardReminder = ({ reminder, title, collection }: DashboardRemi
         }
         initialize()
         debouncedSetIsLoading(true)
-        getData(collectionSchema, constraints, debouncedSetIsLoading, setResults, setUnsubscribe)
+        getData(
+            collectionSchema,
+            constraints,
+            debouncedSetIsLoading,
+            setResultsCallback,
+            setUnsubscribe,
+            reminder.forceServer,
+        )
         return () => {
             unsubscribe?.forEach((unsubscribe) => unsubscribe())
             if (loadingTimeoutRef.current) {
@@ -199,60 +209,26 @@ export const DashboardReminder = ({ reminder, title, collection }: DashboardRemi
                                     </TableRow>
                                 ) : (
                                     results
+                                        .slice(page * pages, (page + 1) * pages)
                                         .sort((a, b) => {
                                             if (!sorting) return 0
                                             const sortingField = getField(fields, sorting.field)
-                                            const valueA = getSortingValue(
-                                                collectionSchema,
+                                            const fieldCustomization = getFieldCustomization(
+                                                sortingField,
                                                 customization,
-                                                sorting.field,
-                                                a,
                                             )
-                                            const valueB = getSortingValue(
-                                                collectionSchema,
-                                                customization,
-                                                sorting.field,
-                                                b,
-                                            )
-                                            let result: number
-                                            if (sortingField.type === "String") {
-                                                const rawA = valueA?.toString().toLowerCase() ?? ""
-                                                const rawB = valueB?.toString().toLowerCase() ?? ""
-                                                result = rawA > rawB ? 1 : rawA < rawB ? -1 : 0
-                                            } else if (isRelationField(sortingField)) {
-                                                const titleField = sortingField.titleField
-                                                let rawA: string
-                                                let rawB: string
-                                                if (titleField) {
-                                                    const recordA = Object.values(a[sortingField.name] ?? {})[0] as
-                                                        | StokerRecord
-                                                        | undefined
-                                                    const recordB = Object.values(b[sortingField.name] ?? {})[0] as
-                                                        | StokerRecord
-                                                        | undefined
-                                                    // eslint-disable-next-line security/detect-object-injection
-                                                    rawA = recordA?.[titleField]?.toString().toLowerCase() ?? ""
-                                                    // eslint-disable-next-line security/detect-object-injection
-                                                    rawB = recordB?.[titleField]?.toString().toLowerCase() ?? ""
-                                                } else {
-                                                    rawA = (
-                                                        Object.keys(a[sortingField.name] ?? {})[0] ?? ""
-                                                    ).toLowerCase()
-                                                    rawB = (
-                                                        Object.keys(b[sortingField.name] ?? {})[0] ?? ""
-                                                    ).toLowerCase()
-                                                }
-                                                result = rawA > rawB ? 1 : rawA < rawB ? -1 : 0
-                                            } else if (sortingField.type === "Timestamp") {
-                                                const rawA = Number(valueA?.valueOf() || 0)
-                                                const rawB = Number(valueB?.valueOf() || 0)
-                                                result = rawA > rawB ? 1 : rawA < rawB ? -1 : 0
-                                            } else {
-                                                result = valueA > valueB ? 1 : valueA < valueB ? -1 : 0
+                                            let sortA = a[sorting.field]
+                                            let sortB = b[sorting.field]
+                                            if (fieldCustomization.admin?.sort) {
+                                                sortA = tryFunction(fieldCustomization.admin?.sort, [a])
+                                                sortB = tryFunction(fieldCustomization.admin?.sort, [b])
                                             }
-                                            return sorting.direction === "asc" ? result : -result
+                                            if (sorting.direction === "asc") {
+                                                return sortA - sortB
+                                            } else {
+                                                return sortB - sortA
+                                            }
                                         })
-                                        .slice(page * pages, (page + 1) * pages)
                                         .map((result: StokerRecord) => (
                                             <TableRow
                                                 key={result.id}