+++ packages/node-client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @stoker-platform/node-client
 
+## 0.5.77
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/types@0.5.56
+    - @stoker-platform/utils@0.5.68
+
 ## 0.5.76
 
 ### Patch Changes